Refactor amphitheater seating into modules
