+++ CHANGELOG.md
@@ -1,9 +1,7 @@
 # Changelog
 
-## 2025-07-30
-- 1100 Expand grass generation to cover the entire world within biome-specific rules, removing the central grass-only circle.
-
 ## 2025-07-29
+- 1357 Break up amphi-seats.js into smaller modules
 - 1013 Make collision detection include players and ignore self-object
 - 0953 Break up trees.js into separate tree modules
 - 1012 Convert global water plane to local pools with 1m recess
